# bigbike-backend/src/main/java/com/bigbike/bigbike_backend/api/customer/CustomerController.java
package com.bigbike.bigbike_backend.api.customer;

import com.bigbike.bigbike_backend.api.common.ApiDataResponse;
import com.bigbike.bigbike_backend.api.common.ApiResponseFactory;
import com.bigbike.bigbike_backend.api.customer.dto.CustomerSummary;
import com.bigbike.bigbike_backend.api.customer.dto.UpdateCustomerProfileRequest;
import com.bigbike.bigbike_backend.api.error.UnauthorizedException;
import com.bigbike.bigbike_backend.domain.customer.CustomerPrincipal;
import com.bigbike.bigbike_backend.service.customer.CustomerAuthService;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

@RestController
@RequestMapping("/api/v1/customer")
@RequiredArgsConstructor
public class CustomerController {

    private final CustomerAuthService authService;
    private final ApiResponseFactory apiResponseFactory;

    @GetMapping("/me")
    public ApiDataResponse<CustomerSummary> me(HttpServletRequest request) {
        CustomerPrincipal principal = requireCustomer();
        return apiResponseFactory.data(authService.getProfile(principal.customerId()), request);
    }

    @PatchMapping("/me")
    public ApiDataResponse<CustomerSummary> updateMe(@Valid @RequestBody UpdateCustomerProfileRequest req, HttpServletRequest request) {
        CustomerPrincipal principal = requireCustomer();
        return apiResponseFactory.data(authService.updateProfile(principal.customerId(), req), request);
    }

    @PostMapping(path = "/me/avatar", consumes = "multipart/form-data")
    public ApiDataResponse<CustomerSummary> uploadAvatar(
            @RequestParam("file") MultipartFile file, HttpServletRequest request) {
        CustomerPrincipal principal = requireCustomer();
        return apiResponseFactory.data(authService.updateAvatar(principal.customerId(), file), request);
    }

    @DeleteMapping("/me/avatar")
    public ApiDataResponse<CustomerSummary> removeAvatar(HttpServletRequest request) {
        CustomerPrincipal principal = requireCustomer();
        return apiResponseFactory.data(authService.removeAvatar(principal.customerId()), request);
    }

    // ── helpers ──────────────────────────────────────────────────────────────

    private CustomerPrincipal requireCustomer() {
        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
        if (auth != null && auth.getPrincipal() instanceof CustomerPrincipal principal) {
            return principal;
        }
        throw new UnauthorizedException("Customer authentication required.");
    }
}