# bigbike-backend/src/main/resources/openapi/bigbike-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BigBike Backend API",
    "version": "1.0.0",
    "description": "REST API for BigBike — Vietnam's premium motorcycle marketplace. Rewrite of WordPress backend. All endpoints return ApiDataResponse<T> or ApiListResponse<T> envelopes."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.bigbike.vn",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/v1/customer/me": {
      "get": {
        "tags": [
          "Customer Profile"
        ],
        "summary": "Get authenticated customer profile",
        "operationId": "getCustomerProfile",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Customer profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSummaryDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/customer/me/avatar": {
      "post": {
        "tags": [
          "Customer Profile"
        ],
        "summary": "Upload/replace own avatar",
        "operationId": "uploadCustomerAvatar",
        "description": "multipart/form-data, single 'file' part. Max 5MB, image/jpeg|png|webp only (declared Content-Type and Tika magic-byte must agree). Replacing an existing avatar deletes the old MinIO object.",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated customer profile with avatarUrl populated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSummaryDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — empty file, oversize, invalid MIME, or content/type mismatch"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Customer Profile"
        ],
        "summary": "Remove own avatar",
        "operationId": "removeCustomerAvatar",
        "description": "Deletes the MinIO object and nulls avatarUrl. Idempotent — calling when already unset is a no-op 200.",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated customer profile with avatarUrl: null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSummaryDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/admin/reviews": {
      "get": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "List reviews (admin)",
        "operationId": "adminListReviews",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
            {
              "name": "rating",
              "in": "query",
              "description": "Optional exact star filter from 1 to 5. The pagination total is filtered; the summary endpoint is not.",
              "schema": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              }
            },
            {
              "name": "lang",
              "in": "query",
              "description": "Content language hint retained for admin compatibility.",
              "schema": {
                "type": "string",
                "default": "vi"
              }
            }
        ],
        "responses": {
          "200": {
            "description": "Filtered review list with pagination.totalItems equal to the filtered result count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/reviews/summary": {
      "get": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Get global review moderation summary (admin)",
        "operationId": "adminReviewSummary",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Global approved rating score/distribution and pending queue counts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewSummaryDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/reviews/{id}": {
      "get": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Get review by id (admin)",
        "operationId": "adminGetReview",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review detail with product metadata, author fields, rating, body, photos, status, and timestamps",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Delete review (admin)",
        "operationId": "adminDeleteReview",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/reviews/{id}/status": {
      "patch": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Update review status (admin)",
        "operationId": "adminUpdateReviewStatus",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated review payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/reviews/bulk-status": {
      "post": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Bulk update review status (admin)",
        "operationId": "adminBulkUpdateReviewStatus",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkReviewStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Number of existing reviews affected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewBulkResultDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/reviews/bulk-delete": {
      "post": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Bulk permanently delete reviews (admin)",
        "operationId": "adminBulkDeleteReviews",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkReviewIdsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Number of existing reviews permanently deleted, including review photos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewBulkResultDataResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT access token. Obtain via POST /api/v1/auth/login. Header: Authorization: Bearer <accessToken>"
      },
      "CustomerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bb_session",
        "description": "Customer session cookie. Set automatically after POST /api/v1/customer/auth/login."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token required for all cart/checkout mutations. Value must match the bb_csrf cookie set by the server."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or token expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "AdminReview": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "productId": { "type": "string", "nullable": true },
          "productName": { "type": "string", "nullable": true },
          "productNameEn": { "type": "string", "nullable": true },
          "productSlug": { "type": "string", "nullable": true },
          "authorName": { "type": "string" },
          "authorEmail": { "type": "string" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "body": { "type": "string" },
          "photos": { "type": "array", "items": { "type": "string" }, "maxItems": 10 },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AdminReviewListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/AdminReview" } },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
      "AdminReviewDataResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdminReview" },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
      "AdminReviewSummary": {
        "type": "object",
        "properties": {
          "approved": { "$ref": "#/components/schemas/AdminReviewApprovedSummary" },
          "pending": { "$ref": "#/components/schemas/AdminReviewPendingSummary" }
        }
      },
      "AdminReviewApprovedSummary": {
        "type": "object",
        "properties": {
          "averageRating": { "type": "number", "format": "double", "minimum": 0, "maximum": 5 },
          "totalReviews": { "type": "integer", "format": "int64" },
          "ratingBreakdown": {
            "type": "object",
            "additionalProperties": { "type": "integer", "format": "int64" },
            "minProperties": 5,
            "maxProperties": 5
          }
        }
      },
      "AdminReviewPendingSummary": {
        "type": "object",
        "properties": {
          "totalReviews": { "type": "integer", "format": "int64" },
          "oneStarReviews": { "type": "integer", "format": "int64" }
        }
      },
      "AdminReviewSummaryDataResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdminReviewSummary" },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
      "UpdateReviewStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": { "status": { "type": "string" } }
      },
      "BulkReviewStatusRequest": {
        "type": "object",
        "required": ["ids", "status"],
        "properties": {
          "ids": { "type": "array", "minItems": 1, "maxItems": 500, "items": { "type": "integer", "format": "int64" } },
          "status": { "type": "string" }
        }
      },
      "BulkReviewIdsRequest": {
        "type": "object",
        "required": ["ids"],
        "properties": {
          "ids": { "type": "array", "minItems": 1, "maxItems": 500, "items": { "type": "integer", "format": "int64" } }
        }
      },
      "AdminReviewBulkResult": {
        "type": "object",
        "properties": { "affected": { "type": "integer", "format": "int32" } }
      },
      "AdminReviewBulkResultDataResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdminReviewBulkResult" },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
...
      "ApiMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
      "CustomerSummary": {
        "type": "object",
        "description": "Customer profile (sensitive fields omitted)",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "newsletterSubscribed": {
            "type": "boolean",
            "description": "Newsletter opt-in flag"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "MinIO object URL (/media/customers/...) for the customer's own uploaded avatar; null when no avatar is set (client renders an initials badge)"
          }
        }
      },
      "CustomerSummaryDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CustomerSummary"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "Standard error envelope used by all error responses",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetail"
            }
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}